Extract TriggerState to new module
